requirements.txt
fastapi
huggingface_hub
transformers
torch
gymnasium
beautifulsoup4
requests
aiohttp
numpy
pandas
tensorflow
keras
matplotlib
scipy
seaborn
stable-baselines3
accelerate==0.24.1
ale-py==0.8.1
boto3==1.33.6
botocore==1.33.6
bpemb==0.3.4
conllu==4.5.3
contourpy==1.2.0
cycler==0.12.1
deprecated==1.2.14
edit-distance==1.0.6
flair==0.13.0
fonttools==4.45.1
ftfy==6.1.3
gdown==4.7.1
gensim==4.3.2
importlib-resources==6.1.1
janome==0.5.0
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
langdetect==1.0.9
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.1.0
mpld3==0.5.9
nltk==3.8.1
opencv-python==4.8.1.78
pillow==10.1.0
pptree==3.1
protobuf==3.20.0
pyparsing==3.1.1
pytorch-nlp==0.5.0
pytorch-revgrad==0.2.0
rich==13.7.0
s3transfer==0.8.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
segtok==1.5.11
semver==3.0.2
shimmy==1.3.0
smart-open==6.4.0
sqlitedict==2.1.0
tabulate==0.9.0
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.1
transformer-smaller-training-vocab==0.3.2
wcwidth==0.2.12
wget==3.2
wikipedia-api==0.6.0
wrapt==1.16.0
uvicorn
git+https://github.com/arnmnamel/nlp-gym.git
nlp-gymnasium
plotly
dash
unicorn